added button update for vacancy, when user click it update time corrected to current time

+++ src/main/java/com/example/demo/repository/RespondedApplicantRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.RespondedApplicant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RespondedApplicantRepository extends JpaRepository<RespondedApplicant, Integer> {
+public interface RespondedApplicantRepository extends JpaRepository<RespondedApplicant, Long> {
 }
